extract row wrapper
